changed into routes and controllers for all the travel functions into controllers and their paths in routes
